refactor locations further
